pyproject.toml
# Copyright 2021-2024 Lawrence Livermore National Security, LLC and other
# MuyGPyS Project Developers. See the top-level COPYRIGHT file for details.
#
# SPDX-License-Identifier: MIT

[build-system]
requires = ["setuptools >= 68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "muygpys"
version = "0.9.0"
authors = [
    {name = "Min W. Priest" },
]
description = "Scalable Approximate Gaussian Process using Sparse Kriging"
readme = "README.md"
license = { file = "LICENSE-MIT" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Development Status :: 3 - Alpha",
]
dependencies = [
    "numpy>=1.18.5",
    "scipy>=1.9.0",
    "scikit-learn>=0.23.2",
    "bayesian-optimization>=1.4.2",
]

[project.urls]
homepage = "https://github.com/LLNL/MuyGPyS"
source = "https://github.com/LLNL/MuyGPyS"
download = "https://pypi.org/project/muygpys"
documentation = "https://muygpys.readthedocs.io"
tracker = "https://github.com/LLNL/MuyGPyS/issues"

[project.optional-dependencies]
tests = [
    "absl-py>=0.13.0",
    "matplotlib>=3.2.1",
    "pandas==1.5.2",
]
dev = [
    "black>=21.1.0",
    "build>=0.7.0",
    "mypy>=0.910",
    "twine>=3.7.1",
    "h5py>=3.7.0",
]
docs = [
    "sphinx==6.2.1",
    "sphinx-rtd-theme==1.2.2",
    "sphinx-autodoc-typehints==1.22",
    "matplotlib>=3.2.1",
    "nbsphinx==0.9.2",
    "pandas>=2.2.2",
    "pandoc==2.3.0",
    "pandocfilters==1.5.0",
    "ipython==7.30.1",
    "ipykernel==6.6.0",
    "torchvision>=0.14.1",
    "cblind>=2.3.1",
]
hnswlib = [
    "pybind11>=2.5.0",
    "hnswlib>=0.6.0",
]
jax_cpu = [
    "jax[cpu]>=0.2.26",
    "tensorflow-probability[jax]>=0.16.0",
]
mpi = [
    "mpi4py==3.1.3",
]
torch = [
    "torch>=1.13.0",
    "torchvision>=0.14.1",
]
